+++ src/backend/metal/src/device.rs
@@ -199,7 +199,7 @@ impl Drop for Device {
 
 bitflags! {
     /// Memory type bits.
-    struct MemoryTypes: u64 {
+    struct MemoryTypes: u32 {
         const PRIVATE = 1<<0;
         const SHARED = 1<<1;
         const MANAGED_UPLOAD = 1<<2;
@@ -1418,7 +1418,7 @@ impl hal::device::Device<Backend> for Device {
                     (Some(&ts.0), Some(&ts.1))
                 } else {
                     (None, None)
-                };    
+                };
 
                 (buffers, attributes, input_assembler, vertex, geometry, hs, ds)
             }
